fix: prevent crash when converting UInt64 to Int in shake animation
- Use modulo on UInt64 before converting to Int to prevent overflow
- Fixes fatal error when state value exceeds Int.max after overflow operations
- Safe conversion: Int(state % UInt64(range)) instead of Int(state) % range

+++ 1.7. Genesis 2 RPG/Features/Battle/BattleView.swift
@@ -776,10 +776,12 @@ private struct ParticipantsPanel: View {
                 var state = seed
                 state = state &* 1103515245 &+ 12345
                 let range = Int(shakeAmplitude * 2) + 1 // Диапазон от -amplitude до +amplitude
-                let xOffset = CGFloat(Int(state) % range) - shakeAmplitude
+                // Safe conversion: use modulo on UInt64 first, then convert to Int
+                let xOffset = CGFloat(Int(state % UInt64(range))) - shakeAmplitude
                 
                 state = state &* 1103515245 &+ 12345
-                let yOffset = CGFloat(Int(state) % range) - shakeAmplitude
+                // Safe conversion: use modulo on UInt64 first, then convert to Int
+                let yOffset = CGFloat(Int(state % UInt64(range))) - shakeAmplitude
                 
                 // Сбрасываем перед новой анимацией
                 shakeX = 0
